Started fixing real alloc, do not check out!
Started fixing the real allocator, it is not working right now. Do not
check out this code! You have been warned :)

+++ Main/Object.cpp
@@ -6,8 +6,6 @@
  */
 
 #include "Object.hpp"
-#include <stdio.h>
-#include <stdlib.h>
 
 namespace traceFileSimulator {
 
@@ -63,7 +61,7 @@ int Object::getGeneration(){
 	return myGeneration;
 }
 
-int Object::getAddress(){
+size_t Object::getAddress(){
 	return myAddress;
 }
 
@@ -110,7 +108,7 @@ int Object::getVisited(){
 	return isVisited;
 }
 
-void Object::updateAddress(int newAddress) {
+void Object::updateAddress(size_t newAddress) {
 	myAddress = newAddress;
 }
